unfrucked the fuckery brought on by reverting the fucked reversions
The problem was when reverting aylins 2 commits, i also deleted the changes she had pulled for those commits.
We reverted the reversions from yesterday and the reversion from the merge of dialogue. Everything is here now.
PROBLEM was in aylins merge that theme ids were changed but not updated in the scene files. we changed them by hand. it now works.